Build fixes Win32.

+++ libs/MAP/CloudMadeMapSource.cpp
@@ -31,22 +31,22 @@ namespace MAP
 	
 	inline static int long2tile( double lon, MagnificationType z )	
 	{ 
-		return (int)( floor( (lon + 180.0) / 360.0 * pow(2.0, z) ) ); 
+		return (int)( floor( (lon + 180.0) / 360.0 * pow(2.0, (double )z) ) ); 
 	}
 	
 	inline static int lat2tile( double lat, MagnificationType z )		
 	{ 
-		return (int)( floor( (1.0 - log( tan( lat * M_PI / 180.0 ) + 1.0 / cos( lat * M_PI/180.0 ) ) / M_PI ) / 2.0 * pow( 2.0, z ) ) ); 
+		return (int)( floor( (1.0 - log( tan( lat * M_PI / 180.0 ) + 1.0 / cos( lat * M_PI/180.0 ) ) / M_PI ) / 2.0 * pow( 2.0, (double)z ) ) ); 
 	}
 	
 	inline static double tile2long( int x, MagnificationType z )		
 	{ 
-		return x / pow( 2.0, z ) * 360.0 - 180; 
+		return x / pow( 2.0, (double)z ) * 360.0 - 180; 
 	}
 	
 	inline static double tile2lat( int y, MagnificationType z )		
 	{ 
-		double n = M_PI - 2.0 * M_PI * y / pow( 2.0, z ); 
+		double n = M_PI - 2.0 * M_PI * y / pow( 2.0, (double)z ); 
 		return 180.0 / M_PI * atan( 0.5 * ( exp( n ) - exp( -n ) ) ); 
 	}
 	
@@ -64,7 +64,7 @@ namespace MAP
 	//-------------------------------------------------------------------------
 	{
 		//return InitialResolution / ( 1 << magnification );
-		return InitialResolution / pow(2.0, magnification);
+		return InitialResolution / pow(2.0, (double)magnification);
 	}
 	
 	//-------------------------------------------------------------------------
